+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-08T09:48:23.645Z",
+    "2026-09-08T14:03:54.711Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {