+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3949,7 +3949,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M206PMHVZ2VNA7ZXKFDDG497"),
+  id: z.string().default("01M20NAGGHXHJA3Y6DTV2XQ292"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>